lib/oelite: Don't include COMPATIBLE_MACHINES and similar in signature
Don't include all the variables used to exclude recipes from being
considered in the recipe metadata signature.  Changes to these should
only be relevant for finding out if the recipe is to be used or not,
and not cause any rebuilds on change.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>

(cherry picked from commit e43b82f809356597b347adac208676f44ac7a064)