lib/oelite: Simplify runq.get_provider() method
* Remove support for PREFERRED_PROVIDER_* variable
* Remove support for PREFERRED_VERSION_* variable
* Set package priority using layer configuration and recipe and package
  preferences as specified by the new 'prefer' command.

This change is not backwards compatible with OE-lite/core 3, as the
the runq.get_provider() method now chooses what to build differently.

The OE-lite/core reference manual will be updated with details on how
the new provider selection scheme works, and how to use it.
A guide for describing how to transition from OE-lite/core 3 should be
given prior to release of OE-lite/core 4.0.0.

The ulstra short version:

When choosing a provider, the package from the top most layer is chosen.
When a recipe is preferred (with a "prefer recipe ..." command), it will
placed in a layer above highest normal metadata layer.
When a package is preferred (with a "prefer package ..." command), it will
be placed in a layer above the layer for preferred recipes.
When choosing among providers in the same layer, the package with the
highest PRIORITY will be chosen (package specific priorities can be set
using PRIORITY_name-of-package style variables).
When choosing among multiple provider with the same priority in the same
layer, the newest version will be chosen.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>