class/autotools: Fix corner-case for CONFIGURE_SCRIPT variable
If (when) a recipe is machine specific, the weak assignment to
CONFIGURE_SCRIPT is causing problems.  It ends up assigning the
generic (not machine specific) WORKDIR value instead of the correct
machine specific variant, which is created by a hook later on.

This commit works around it, using an extra variable. *sigh*

Signed-off-by: Esben Haabendal <esben@haabendal.dk>